+++ spec/features/steps/image_steps.rb
@@ -1,4 +1,4 @@
-step 'a/an image for image series :image_series_instance' do |image_series|
+step 'a/an (DICOM )image for image series :image_series_instance' do |image_series|
   @image = create(:image, image_series: image_series)
 end