+++ spec/features/steps/placeholders.rb
@@ -1,3 +1,28 @@
+class TurnipHelper
+  class << self
+    def find_record(type, identifier)
+      record =
+        case type
+        when 'BackgroundJob' then BackgroundJob.find_by(id: identifier)
+        when 'Study' then Study.find_by(name: identifier)
+        when 'Center' then Center.find_by(name: identifier)
+        when 'Patient' then Patient.find_by(subject_id: identifier)
+        when 'Visit' then Visit.find_by(visit_number: identifier)
+        when 'ImageSeries' then ImageSeries.find_by(name: identifier)
+        when 'Image' then
+          image_series_identifier, image_index = identifier.split("#")
+          image_series = ImageSeries.find_by(name: image_series_identifier)
+          image_series.images.order(id: :asc)[image_index.to_i - 1]
+        when 'User' then User.find_by(username: identifier)
+        when 'Role' then Role.find_by(title: identifier)
+        end
+      fail "Cannot find #{type} with identifier #{identifier}" if record.nil?
+
+      record
+    end
+  end
+end
+
 placeholder :activity do
   match(/([^ $\n]+)/) do |activity|
     activities = Ability::ACTIVITIES.values.flatten.uniq
@@ -36,40 +61,18 @@ placeholder :admin_path do
       .send("admin_#{model_name.singularize.underscore}_path", record)
   end
 
-  match(/([^ $\n]+) ([^ $]+) "([^$\n]+)"/) do |action, model_name, identifier|
+  match(/([^ $\n]+) (for )?([^ $]+) "([^$\n]+)"/) do |action, _, model_name, identifier|
     method = "admin_#{model_name.underscore}_path"
     method = "#{action}_#{method}" if action != 'show'
     model_name = model_name.classify
-    record =
-      case model_name
-      when 'BackgroundJob' then BackgroundJob.find_by(id: identifier)
-      when 'Study' then Study.find_by(name: identifier)
-      when 'Center' then Center.find_by(name: identifier)
-      when 'Patient' then Patient.find_by(subject_id: identifier)
-      when 'Visit' then Visit.find_by(visit_number: identifier)
-      when 'ImageSeries' then ImageSeries.find_by(name: identifier)
-      when 'Image' then Image.find_by(id: identifier)
-      when 'User' then User.find_by(username: identifier)
-      when 'Role' then Role.find_by(title: identifier)
-      end
+    record = TurnipHelper.find_record(model_name, identifier)
     Rails.application.routes.url_helpers
          .send(method, record)
   end
 
   match(/([^ $]+) "([^$\n]+)"/) do |model_name, identifier|
     model_name = model_name.classify
-    record =
-      case model_name
-      when 'BackgroundJob' then BackgroundJob.find_by(id: identifier)
-      when 'Study' then Study.find_by(name: identifier)
-      when 'Center' then Center.find_by(name: identifier)
-      when 'Patient' then Patient.find_by(subject_id: identifier)
-      when 'Visit' then Visit.find_by(visit_number: identifier)
-      when 'ImageSeries' then ImageSeries.find_by(name: identifier)
-      when 'Image' then Image.find_by(id: identifier)
-      when 'User' then User.find_by(username: identifier)
-      when 'Role' then Role.find_by(title: identifier)
-      end
+    record = TurnipHelper.find_record(model_name, identifier)
     Rails.application.routes.url_helpers
          .send("admin_#{model_name.singularize.underscore}_path", record)
   end
@@ -94,15 +97,7 @@ end
 placeholder :model_instance do
   match(/([^ $]+) "([^$\n]+)"/) do |model_name, identifier|
     model_name = model_name.classify
-    case model_name
-    when 'BackgroundJob' then BackgroundJob.find_by(id: identifier)
-    when 'Study' then Study.find_by(name: identifier)
-    when 'Center' then Center.find_by(name: identifier)
-    when 'Patient' then Patient.find_by(subject_id: identifier)
-    when 'User' then User.find_by(username: identifier)
-    when 'Visit' then Visit.find_by(visit_number: identifier)
-    when 'ImageSeries' then ImageSeries.find_by(name: identifier)
-    end
+    TurnipHelper.find_record(model_name, identifier)
   end
 end