+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Minor Changes
 
 * Feature - Allow cleaning DICOM tag values for all images of a whole image series or a whole study (depending on authorization)
+* Feature - Display nested DICOM attributes and `SQ` VR accordingly
+* Permissions - Manage access to DICOM metadata via the new `read_dicom_metadata` permission for `ImageSeries`
 
 ### DevOps